requirements.txt
feedparser==6.0.11
python-dotenv==1.0.1
Jinja2==3.1.4
requests==2.32.3
